+++ SJCNet.DesignPatterns.Factory.OLD/Properties/AssemblyInfo.cs
