+++ SJCNet.DesignPatterns.Factory.OLD/Shared/Automobile.cs
@@ -1,6 +1,6 @@
 ﻿using SJCNet.DesignPatterns.Shared.Utility.Old;
 
-namespace SJCNet.DesignPatterns.Factory.Shared
+namespace SJCNet.DesignPatterns.Factory.Old.Shared
 {
     public abstract class Automobile : IAutomobile
     {
