+++ SJCNet.DesignPatterns.Factory.OLD/Shared/Car.cs
@@ -1,4 +1,4 @@
-﻿namespace SJCNet.DesignPatterns.Factory.Shared
+﻿namespace SJCNet.DesignPatterns.Factory.Old.Shared
 {
     public class Car : Automobile, ICar
     {
