+++ SJCNet.DesignPatterns.Factory.OLD/Shared/ICar.cs
@@ -1,4 +1,4 @@
-﻿namespace SJCNet.DesignPatterns.Factory.Shared
+﻿namespace SJCNet.DesignPatterns.Factory.Old.Shared
 {
     public interface ICar : IAutomobile
     {
