+++ SJCNet.DesignPatterns.Factory.OLD/Shared/IMotorbike.cs
@@ -1,4 +1,4 @@
-﻿namespace SJCNet.DesignPatterns.Factory.Shared
+﻿namespace SJCNet.DesignPatterns.Factory.Old.Shared
 {
     public interface IMotorbike : IAutomobile
     {
