+++ SJCNet.DesignPatterns.Factory.OLD/Shared/IVan.cs
@@ -1,4 +1,4 @@
-﻿namespace SJCNet.DesignPatterns.Factory.Shared
+﻿namespace SJCNet.DesignPatterns.Factory.Old.Shared
 {
     public interface IVan : IAutomobile
     {
