+++ SJCNet.DesignPatterns.Output.OLD/Program.cs
@@ -18,8 +18,8 @@ namespace SJCNet.DesignPatterns.Output.Old
             //observerPattern.Execute();
 
             // Example: Factory Pattern
-            var factoryPattern = new Factory.Example();
-            factoryPattern.Execute();
+            //var factoryPattern = new Factory.Example();
+            //factoryPattern.Execute();
 
             Logger.WriteEnd("Ended");
             ReadLine();
