+++ SJCNet.DesignPatterns.Output.OLD/SJCNet.DesignPatterns.Output.OLD.csproj
@@ -66,10 +66,6 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SJCNet.DesignPatterns.Factory\SJCNet.DesignPatterns.Factory.csproj">
-      <Project>{5905da6d-16c1-4dde-9c78-552e59055e7c}</Project>
-      <Name>SJCNet.DesignPatterns.Factory</Name>
-    </ProjectReference>
     <ProjectReference Include="..\SJCNet.DesignPatterns.Observer\SJCNet.DesignPatterns.Observer.csproj">
       <Project>{fba2123e-a1c9-4327-aa17-33f35481964a}</Project>
       <Name>SJCNet.DesignPatterns.Observer</Name>
